+++ pkg/appmgmt/general/general.go
@@ -20,7 +20,7 @@ package general
 
 import (
 	"go.uber.org/zap"
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/selection"
 	k8sCache "k8s.io/client-go/tools/cache"
@@ -88,10 +88,13 @@ func (os *Manager) getTaskMetadata(pod *v1.Pod) (interfaces.TaskMetadata, bool) 
 		return interfaces.TaskMetadata{}, false
 	}
 
+	taskGroupName := utils.GetTaskGroupFromPodSpec(pod)
+
 	return interfaces.TaskMetadata{
 		ApplicationID: appId,
 		TaskID:        string(pod.UID),
 		Pod:           pod,
+		TaskGroupName: taskGroupName,
 	}, true
 }
 
@@ -114,11 +117,17 @@ func (os *Manager) getAppMetadata(pod *v1.Pod) (interfaces.ApplicationMetadata, 
 	// get the application owner (this is all that is available as far as we can find)
 	user := pod.Spec.ServiceAccountName
 
+	taskGroups, err := utils.GetTaskGroupsFromAnnotation(pod)
+	if err != nil {
+		log.Logger().Error("unable to get taskGroups by given pod", zap.Error(err))
+	}
+
 	return interfaces.ApplicationMetadata{
 		ApplicationID: appId,
 		QueueName:     utils.GetQueueNameFromPod(pod),
 		User:          user,
 		Tags:          tags,
+		TaskGroups:    taskGroups,
 	}, true
 }
